moved sync ahead of gui tests
