+++ src/test/java/address/async/cloud/CloudSimulatorTest.java
@@ -1,6 +1,10 @@
-package address.sync.cloud;
+package address.async.cloud;
 
 import address.exceptions.DataConversionException;
+import address.sync.cloud.CloudFileHandler;
+import address.sync.cloud.CloudRateLimitStatus;
+import address.sync.cloud.CloudSimulator;
+import address.sync.cloud.RemoteResponse;
 import address.sync.cloud.model.CloudAddressBook;
 import address.sync.cloud.model.CloudPerson;
 import address.sync.cloud.model.CloudTag;
